backend/requirements.txt
# Bank Teller Chatbot - Optimized Dependencies
# Updated: 2024 - Latest stable versions

# ============================================================================
# Core Framework
# ============================================================================
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
python-dotenv==1.0.1

# ============================================================================
# Machine Learning - LIGHTWEIGHT OPTION (RECOMMENDED)
# ============================================================================
# Use tensorflow-cpu for smaller size and faster install (~200MB vs ~500MB)
tensorflow-cpu==2.17.0
# OR if you need GPU support, uncomment below and comment above:
# tensorflow==2.17.0

scikit-learn==1.5.1

# ============================================================================
# NLP - OPTIONAL (Start without, add if needed)
# ============================================================================
# spacy==3.7.5
# To download model later: python -m spacy download en_core_web_sm
# NOTE: We'll use regex-first approach for entity extraction

# ============================================================================
# Data Processing
# ============================================================================
pandas==2.2.2
numpy==1.26.4
datasets==2.20.0

# ============================================================================
# Validation & Models
# ============================================================================
pydantic==2.8.2
pydantic-settings==2.4.0
bleach==6.1.0

# ============================================================================
# Database - Async Support
# ============================================================================
aiosqlite==0.20.0

# ============================================================================
# Testing
# ============================================================================
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# ============================================================================
# Security & Utilities
# ============================================================================
bleach==6.1.0           # XSS/HTML sanitization
python-dateutil==2.9.0

# ============================================================================
# Development Tools (Optional - uncomment if needed)
# ============================================================================
# black==24.4.2          # Code formatting
# flake8==7.1.0          # Linting
# mypy==1.10.1           # Type checking
# ipython==8.26.0        # Better REPL